requirements.txt
requests>=2.25.1
python-telegram-bot==20.5
